requirements.txt
# 웹앱 필수 패키지
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0

# 엑셀 처리
openpyxl==3.1.2
python-dateutil==2.8.2

# 데이터 검증
marshmallow==3.20.1

# AWS 배포용 (선택사항)
gunicorn==21.2.0
boto3==1.34.0

# 개발용
pytest==7.4.3

# Streamlit UI
streamlit==1.41.1
streamlit-aggrid==0.3.6
pandas==2.2.3





